requirements.txt
Flask==2.0.2
openai==0.27.0
pinecone-client==2.0.0
pdfplumber==0.5.28
langchain==0.0.40
python-dotenv==0.21.0
langchain-pinecone==0.1.3